[test] Restore existing behavior of the Papaya diagram description
