+++ packages/sirius-web/backend/sirius-web-sample-application/src/main/java/org/eclipse/sirius/web/sample/papaya/view/logicalarchitecture/RealizedByEdgeDescriptionProvider.java
@@ -1,0 +1,74 @@
+/*******************************************************************************
+ * Copyright (c) 2023 Obeo.
+ * This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v2.0
+ * which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     Obeo - initial API and implementation
+ *******************************************************************************/
+package org.eclipse.sirius.web.sample.papaya.view.logicalarchitecture;
+
+import org.eclipse.sirius.components.view.ArrowStyle;
+import org.eclipse.sirius.components.view.DiagramDescription;
+import org.eclipse.sirius.components.view.EdgeDescription;
+import org.eclipse.sirius.components.view.LineStyle;
+import org.eclipse.sirius.components.view.ViewFactory;
+import org.eclipse.sirius.web.sample.papaya.view.IEdgeDescriptionProvider;
+import org.eclipse.sirius.web.sample.papaya.view.PapayaViewCache;
+
+/**
+ * Description of realized by.
+ *
+ * @author sbegaudeau
+ */
+public class RealizedByEdgeDescriptionProvider implements IEdgeDescriptionProvider {
+
+    @Override
+    public EdgeDescription create() {
+        var realizedByEdgeStyle = ViewFactory.eINSTANCE.createEdgeStyle();
+        realizedByEdgeStyle.setColor("#fb8c00");
+        realizedByEdgeStyle.setEdgeWidth(1);
+        realizedByEdgeStyle.setLineStyle(LineStyle.SOLID);
+        realizedByEdgeStyle.setSourceArrowStyle(ArrowStyle.NONE);
+        realizedByEdgeStyle.setTargetArrowStyle(ArrowStyle.INPUT_ARROW);
+
+        var realizedByEdgeDescription = ViewFactory.eINSTANCE.createEdgeDescription();
+        realizedByEdgeDescription.setName("Edge Realized by");
+        realizedByEdgeDescription.setLabelExpression("aql:self.name + ' is realized by ' + self.realizedBy.name");
+        realizedByEdgeDescription.setStyle(realizedByEdgeStyle);
+        realizedByEdgeDescription.setSourceNodesExpression("aql:self");
+        realizedByEdgeDescription.setTargetNodesExpression("aql:self.realizedBy");
+        realizedByEdgeDescription.setIsDomainBasedEdge(false);
+
+        var realizedByEdgeTool = ViewFactory.eINSTANCE.createEdgeTool();
+        realizedByEdgeTool.setName("Realized by");
+        var changeContext = ViewFactory.eINSTANCE.createChangeContext();
+        changeContext.setExpression("aql:semanticEdgeSource");
+        var setTargetValue = ViewFactory.eINSTANCE.createSetValue();
+        setTargetValue.setFeatureName("realizedBy");
+        setTargetValue.setValueExpression("aql:semanticEdgeTarget");
+
+        changeContext.getChildren().add(setTargetValue);
+        realizedByEdgeTool.getBody().add(changeContext);
+
+        realizedByEdgeDescription.getEdgeTools().add(realizedByEdgeTool);
+
+        return realizedByEdgeDescription;
+    }
+
+    @Override
+    public void link(DiagramDescription diagramDescription, PapayaViewCache cache) {
+        var realizedByEdgeDescription = cache.getEdgeDescription("Edge Realized by");
+        var operationalActivityNodeDescription = cache.getNodeDescription("Node papaya::OperationalActivity");
+        var componentNodeDescription = cache.getNodeDescription("Node papaya::Component");
+
+        diagramDescription.getEdgeDescriptions().add(realizedByEdgeDescription);
+        realizedByEdgeDescription.getSourceNodeDescriptions().add(operationalActivityNodeDescription);
+        realizedByEdgeDescription.getTargetNodeDescriptions().add(componentNodeDescription);
+    }
+
+}
